rename package ubik to ubikagent
